release: 0.1.0-alpha.17

+++ src/stigg/resources/v1/coupons.py
@@ -57,7 +57,7 @@ class CouponsResource(SyncAPIResource):
         duration_in_months: Optional[int],
         metadata: Optional[Dict[str, str]],
         name: str,
-        percent_off: Optional[float],
+        percent_off: Optional[int],
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
         extra_headers: Headers | None = None,
@@ -327,7 +327,7 @@ class AsyncCouponsResource(AsyncAPIResource):
         duration_in_months: Optional[int],
         metadata: Optional[Dict[str, str]],
         name: str,
-        percent_off: Optional[float],
+        percent_off: Optional[int],
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
         extra_headers: Headers | None = None,
